Title: 최솟값 3개, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 250207/최솟값 3개/top-3-smallest-number.cpp
@@ -21,7 +21,7 @@ int main() {
             a= pq.top(); pq.pop();
             b= pq.top(); pq.pop();
             c= pq.top(); pq.pop();
-            result = a*b*c;
+            result = long(a*b*c);
             pq.push(a);pq.push(b);pq.push(c);
             cout << result<<endl;
         }
